requirements.txt
Flask==2.2.2
GitPython==3.1.29
Jinja2==3.1.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
Pillow==8.2.0
PyJWT==2.6.0
PyNaCl==1.5.0
PySocks==1.7.1
PyWavelets==1.3.0
PyYAML==6.0
Pygments==2.13.0
QtPy==2.3.0
SQLAlchemy==1.4.45
SecretStorage==3.3.3
Send2Trash==1.8.0
Werkzeug==2.2.2
absl-py==1.3.0
adal==1.2.7
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
anyio==3.6.2
argcomplete==2.0.0
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
azure-common==1.1.28
azure-core==1.26.1
azure-graphrbac==0.61.1
azure-mgmt-authorization==3.0.0
azure-mgmt-containerregistry==10.0.0
azure-mgmt-core==1.3.2
azure-mgmt-keyvault==10.1.0
azure-mgmt-resource==21.2.1
azure-mgmt-storage==20.1.0
azureml-core==1.48.0
azureml==0.2.7
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
contextlib2==21.6.0
cryptography==38.0.4
cycler==0.11.0
databricks-cli==0.17.4
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
docker==6.0.1
entrypoints==0.4
environs==9.3.3
fastjsonschema==2.16.2
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
pycrumbs==0.1.3
gitdb==4.0.10
google-auth-oauthlib==0.4.6
google-auth==1.35.0
greenlet==2.0.1
grpcio==1.51.1
gunicorn==20.1.0
humanfriendly==10.0
idna==3.4
imageio==2.9.0
importlib-metadata==4.13.0
importlib-resources==5.10.1
ipykernel==6.3.1
ipython-genutils==0.2.0
ipython==7.34.0
ipywidgets==8.0.3
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.8.0
jmespath==1.0.1
joblib==1.2.0
jsonpickle==2.2.0
jsonschema==4.17.3
jupyter-client==6.1.5
jupyter-console==6.4.2
jupyter-server==1.15.6
jupyter==1.0.0
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
kiwisolver==1.4.4
knack==0.10.1
lightning-bolts==0.3.4
lightning-utilities==0.4.2
marshmallow==3.19.0
matplotlib-inline==0.1.6
matplotlib==3.2.1
mistune==2.0.4
mlflow==1.30.0
msal-extensions==1.0.0
msal==1.20.0
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.3
mypy-extensions==0.4.3
mypy==0.812
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.6
nbformat==5.7.0
ndg-httpsclient==0.5.1
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.19.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
param==1.9.3
paramiko==2.12.0
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.9.2
pkgutil_resolve_name==1.3.10
plotly==5.3.1
plotnine==0.8.0
portalocker==2.6.0
prometheus-client==0.15.0
prometheus-flask-exporter==0.21.0
prompt-toolkit==3.0.36
protobuf==3.20.1
ptyprocess==0.7.0
pyDeprecate==0.3.1
pyOpenSSL==22.1.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pydicom==2.3.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-gdcm==3.0.20
pytorch-lightning==1.4.7
pytz==2022.6
pyzmq==24.0.1
qtconsole==5.4.0
querystring-parser==1.2.4
requests-oauthlib==1.3.1
requests==2.28.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.2
six==1.16.0
sklearn-pandas==2.2.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tabulate==0.8.7
tenacity==8.1.0
tensorboard-plugin-wit==1.8.1
tensorboard==2.3.0
tensorboardX==2.5.1
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
torch==1.13.0
torchmetrics==0.4.1
torchprof==1.3.3
torchvision==0.14.0
tornado==6.2
tqdm==4.59.0
traitlets==5.7.1
typed-ast==1.4.3
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.4
yarl==1.8.2
zipp==3.11.0
monai==1.1.0